services/payments/requirements.txt
flask==2.0.3
stripe
firebase-admin
kafka-python
flask-cors